Create economicclass.c
Problem
Alice has recently learned in her economics class that the market is said to be in equilibrium when the supply is equal to the demand.

Alice has market data for N time points in the form of two integer arrays S and D. 
Help Alice in finding out the number of time points at which the market was in equilibrium.

Input Format
The first line of input will contain an integer T— the number of test cases. 
The description of T test cases follows.
Each test case contains three lines of input.
The first line of each test case contains an integer NN, the length of the arrays SS and DD.
The second line of each test case contains NN space-separated integers S1, S2,...., SN.
The third line of each test case contains N space-separated integers D1, D2,...., DN.

Output Format
For each test case, output the number of time points at which the market was in equilibrium.